refactor(vector-search): move vector search functions to a separate module
