+++ src/lib/rag.mjs
@@ -1,17 +1,17 @@
-import { ChatOpenAI, OpenAIEmbeddings } from '@langchain/openai';
-import { MongoClient } from 'mongodb';
-import { MONGODB_URI, OPENAI_API_KEY, EMBEDDING_MODEL, VECTOR_INDEX_NAME } from './vars.mjs';
+import { ChatOpenAI } from '@langchain/openai';
+import { OPENAI_API_KEY } from './vars.mjs';
 import * as z from 'zod';
 import fs from 'node:fs';
 import path from 'node:path';
+import { vectorSearch } from './vector-search.mjs';
+export { closeConnection } from './vector-search.mjs';
 
 const isProduction = process.env.NODE_ENV === 'production';
 
 const gpt5Nano = new ChatOpenAI({
   modelName: 'gpt-5-nano',
-  maxTokens: 1000,
   apiKey: OPENAI_API_KEY,
-  temperature: 0.1,
+  temperature: 1,
 });
 
 const gpt4oMini = new ChatOpenAI({
@@ -20,11 +20,6 @@ const gpt4oMini = new ChatOpenAI({
   apiKey: OPENAI_API_KEY,
   temperature: 0.1,
 });
-
-const DB_NAME = 'IRCC_RAG';
-const COLLECTION_NAME = 'chunks';
-const mongoClient = new MongoClient(MONGODB_URI, {});
-const collection = mongoClient.db(DB_NAME).collection(COLLECTION_NAME);
 
 export async function rewriteQuery(query, messageHistory = [], model = gpt5Nano) {
   const rewriteQueryPrompt = `You are an assistant that reformulates user questions to improve information retrieval.
@@ -133,151 +128,10 @@ async function logInteractionResults(ragPrompt, responseObject) {
   fs.writeFileSync(logFileName, logContent, 'utf8');
 }
 
-export async function vectorSearch(query) {
-  const embeddings = await new OpenAIEmbeddings({
-    openAIApiKey: OPENAI_API_KEY,
-    model: EMBEDDING_MODEL,
-  }).embedQuery(query);
-
-  await openMongoConnection();
-
-  const SEPARATOR = '\n\n\\[...\\]\n\n* * * \n\n';
-
-  const results = await collection
-    .aggregate([
-      {
-        // Any change to these parameters alters the amount of results passed to the LLM,
-        // thus change the context and the quality of the answer.
-        $vectorSearch: {
-          queryVector: embeddings,
-          path: 'embedding',
-          numCandidates: 500,
-          index: VECTOR_INDEX_NAME,
-          limit: 20,
-        },
-      },
-      {
-        $sort: {
-          refUrl: 1,
-          'loc.lines.from': 1,
-        },
-      },
-      {
-        $group: {
-          _id: '$refUrl',
-          minLoc: {
-            $min: '$loc.lines.from',
-          },
-          text: {
-            $push: '$text',
-          },
-        },
-      },
-      // Context reconstruction:
-      // Add the first chunk of the related document as the main topic.
-      {
-        $lookup: {
-          from: 'chunks',
-          let: {
-            refUrl: '$_id',
-            minLoc: '$minLoc',
-          },
-          pipeline: [
-            {
-              $match: {
-                $expr: {
-                  $and: [
-                    {
-                      $eq: ['$refUrl', '$$refUrl'],
-                    },
-                    {
-                      $ne: ['$loc.lines.from', '$$minLoc'],
-                    },
-                  ],
-                },
-              },
-            },
-            {
-              $sort: {
-                'loc.lines.from': 1,
-              },
-            },
-            {
-              $limit: 1,
-            },
-          ],
-          as: 'relatedDocs',
-        },
-      },
-      {
-        $addFields: {
-          relatedDocs: {
-            $arrayElemAt: ['$relatedDocs', 0],
-          },
-          text: {
-            $reduce: {
-              input: '$text',
-              initialValue: '',
-              in: {
-                $concat: ['$$value', '$$this', SEPARATOR],
-              },
-            },
-          },
-        },
-      },
-      {
-        $project: {
-          _id: 0,
-          refUrl: '$_id',
-          text: 1,
-          mainTopic: '$relatedDocs.text',
-        },
-      },
-    ])
-    .toArray();
-
-  return results;
-}
-
-/**
- * Converts an array of chunks into a Markdown-formatted string.
- *
- * @param {Array<Object>} chunks - An array of chunk objects to be converted.
- * @param {string} chunks[].refUrl - The reference URL associated with the chunk.
- * @param {string} chunks[].text - The text content of the chunk.
- * @param {string?} chunks[].mainTopic - The initial text chunk in the related full document.
- * @returns {string} A Markdown-formatted string containing the chunk text and references.
- */
-export function chunksToMarkdown(chunks) {
-  return chunks
-    .map(({ refUrl, text, mainTopic }, i) => {
-      // TO-DO: implement text fragment highlight
-      // const firstSentence = removeMd(text).match(/[\w, -]{12,}/).at(0);
-      // const lastSentence = removeMd(text).match(/[\w, -]{12,}/g).at(-1);
-      // let url = refUrl;
-      // if (firstSentence && lastSentence && firstSentence !== lastSentence) {
-      //   url = `${refUrl}#:~:text=${encodeURI(firstSentence)},${encodeURI(lastSentence)}`;
-      // }
-      return `${mainTopic ? mainTopic + SEPARATOR : ''}${text}\n\nReference [${
-        i + 1
-      }]: ${refUrl}\n-----------------------------\n\n`;
-    })
-    .join('\n');
-}
-
-async function openMongoConnection() {
-  await mongoClient.connect();
-  return mongoClient;
-}
-
 export async function ask(query, messageHistory = []) {
   const response = await rewriteQuery(query, messageHistory);
   if (response.error) {
     return response;
   }
   return RAG(response.answer);
-}
-
-export async function closeConnection() {
-  await mongoClient?.close();
 }
